# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for comparing API responses and detecting differences",
        "title": "Checker API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/compare": {
            "post": {
                "description": "Compares two API responses and returns the differences between them",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comparisons"
                ],
                "summary": "Compare two API responses",
                "parameters": [
                    {
                        "description": "Comparison request with source and destination responses",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CompareRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comparison result",
                        "schema": {
                            "$ref": "#/definitions/model.CompareResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/compare/{id}": {
            "get": {
                "description": "Retrieves a comparison from the repository by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comparisons"
                ],
                "summary": "Get a comparison by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comparison ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comparison result",
                        "schema": {
                            "$ref": "#/definitions/model.CompareResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Comparison not found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.CompareRequest": {
            "description": "Request payload for comparing two API responses",
            "type": "object",
            "properties": {
                "destination": {
                    "description": "Destination response to compare against",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Response"
                        }
                    ]
                },
                "source": {
                    "description": "Source response to compare",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Response"
                        }
                    ]
                }
            }
        },
        "model.CompareResponse": {
            "description": "Response containing the comparison result with differences",
            "type": "object",
            "properties": {
                "differences": {
                    "description": "List of differences found between the responses",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Difference"
                    }
                },
                "id": {
                    "description": "Unique identifier for the comparison",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "is_equal": {
                    "description": "Whether the responses are equal",
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "model.Difference": {
            "description": "Details about a specific difference found between responses",
            "type": "object",
            "properties": {
                "description": {
                    "description": "Human-readable description of the difference",
                    "type": "string",
                    "example": "Status codes differ: 200 vs 404"
                },
                "dest_value": {
                    "description": "Value in the destination response (can be any type: string, number, boolean, etc.)"
                },
                "path": {
                    "description": "Path to the difference (e.g., \"body.user.name\", \"headers.Content-Type\")",
                    "type": "string",
                    "example": "status"
                },
                "source_value": {
                    "description": "Value in the source response (can be any type: string, number, boolean, etc.)"
                },
                "type": {
                    "description": "Type of difference (status, header, body, missing, extra)",
                    "type": "string",
                    "example": "status"
                }
            }
        },
        "model.ErrorResponse": {
            "description": "Error response structure",
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error type or code",
                    "type": "string",
                    "example": "Invalid request body"
                },
                "message": {
                    "description": "Detailed error message",
                    "type": "string",
                    "example": "Failed to parse JSON"
                }
            }
        },
        "model.Response": {
            "description": "API response structure containing status, headers, and body",
            "type": "object",
            "properties": {
                "body": {
                    "description": "Response body as bytes (base64 encoded in JSON)",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "headers": {
                    "description": "HTTP headers",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "status": {
                    "description": "HTTP status code",
                    "type": "integer"
                }
            }
        }
    }
}